requirements.txt
docker==6.0.1
nvidia-ml-py==11.450.129
prometheus-client==0.16.0
pydantic==1.10.7
PyYAML==6.0
